# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Industrial Safety Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RiskLevel = typeof RiskLevel[keyof typeof RiskLevel];


export const RiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export type HazardStatus = typeof HazardStatus[keyof typeof HazardStatus];


export const HazardStatus = {
  active: 'active',
  acknowledged: 'acknowledged',
  monitoring: 'monitoring',
  resolved: 'resolved',
} as const;

export type SensorType = typeof SensorType[keyof typeof SensorType];


export const SensorType = {
  gas: 'gas',
  vibration: 'vibration',
  temperature: 'temperature',
  pressure: 'pressure',
  proximity: 'proximity',
  camera: 'camera',
} as const;

export type SensorStatus = typeof SensorStatus[keyof typeof SensorStatus];


export const SensorStatus = {
  normal: 'normal',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface Zone {
  id: string;
  name: string;
  riskLevel: RiskLevel;
  activeHazardCount: number;
  workersPresent: number;
  activeWorkPermits: number;
}

export interface Sensor {
  id: string;
  name: string;
  sensorType: SensorType;
  zoneId: string;
  zoneName: string;
  unit: string;
  latestValue: number;
  status: SensorStatus;
  updatedAt: string;
}

export interface SensorReading {
  timestamp: string;
  value: number;
  status: SensorStatus;
}

export interface ContributingFactor {
  id: string;
  label: string;
  description: string;
  /** Relative contribution to the overall risk assessment, 0-1 */
  weight: number;
}

export type RecommendedActionPriority = typeof RecommendedActionPriority[keyof typeof RecommendedActionPriority];


export const RecommendedActionPriority = {
  immediate: 'immediate',
  high: 'high',
  normal: 'normal',
} as const;

export interface RecommendedAction {
  id: string;
  action: string;
  rationale: string;
  priority: RecommendedActionPriority;
}

export interface ExplainabilityTrace {
  contributingSensors: string[];
  reasoningSteps: string[];
  riskRationale: string;
  priorityRationale: string;
  confidenceRationale: string;
}

export interface HazardSummary {
  id: string;
  title: string;
  scenarioType: string;
  riskLevel: RiskLevel;
  confidenceScore: number;
  priorityScore: number;
  status: HazardStatus;
  zoneId: string;
  zoneName: string;
  detectedAt: string;
  workersNearby: number;
  activeWorkPermit: boolean;
}

export type HazardDetail = HazardSummary & {
  description: string;
  contributingFactors: ContributingFactor[];
  recommendedActions: RecommendedAction[];
  explainability: ExplainabilityTrace;
};

export interface HazardStatusUpdate {
  status: HazardStatus;
}

export interface DashboardSummary {
  activeHazardCount: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  averageConfidence: number;
  zonesMonitored: number;
  sensorsOnline: number;
  workersOnSite: number;
  topHazards: HazardSummary[];
}

export type TimelineEventEventType = typeof TimelineEventEventType[keyof typeof TimelineEventEventType];


export const TimelineEventEventType = {
  sensor_anomaly: 'sensor_anomaly',
  scenario_detected: 'scenario_detected',
  risk_escalated: 'risk_escalated',
  priority_changed: 'priority_changed',
  recommendation_issued: 'recommendation_issued',
  status_changed: 'status_changed',
} as const;

export interface TimelineEvent {
  id: string;
  hazardId: string;
  hazardTitle: string;
  eventType: TimelineEventEventType;
  description: string;
  timestamp: string;
}

export type ListSensorsParams = {
zoneId?: string;
sensorType?: SensorType;
};

export type ListHazardsParams = {
status?: HazardStatus;
riskLevel?: RiskLevel;
};

export type ListActivityParams = {
limit?: number;
};