# lib/api-zod/src/generated/types/contributingFactor.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Industrial Safety Intelligence System API
 * OpenAPI spec version: 0.1.0
 */

export interface ContributingFactor {
  id: string;
  label: string;
  description: string;
  /** Relative contribution to the overall risk assessment, 0-1 */
  weight: number;
}