# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Industrial Safety Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { HazardSummary } from './hazardSummary';

export interface DashboardSummary {
  activeHazardCount: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  averageConfidence: number;
  zonesMonitored: number;
  sensorsOnline: number;
  workersOnSite: number;
  topHazards: HazardSummary[];
}