# lib/api-zod/src/generated/types/explainabilityTrace.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Industrial Safety Intelligence System API
 * OpenAPI spec version: 0.1.0
 */

export interface ExplainabilityTrace {
  contributingSensors: string[];
  reasoningSteps: string[];
  riskRationale: string;
  priorityRationale: string;
  confidenceRationale: string;
}