# lib/api-zod/src/generated/types/hazardDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Industrial Safety Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { ContributingFactor } from './contributingFactor';
import type { ExplainabilityTrace } from './explainabilityTrace';
import type { HazardSummary } from './hazardSummary';
import type { RecommendedAction } from './recommendedAction';

export type HazardDetail = HazardSummary & {
  description: string;
  contributingFactors: ContributingFactor[];
  recommendedActions: RecommendedAction[];
  explainability: ExplainabilityTrace;
};