# lib/api-zod/src/generated/types/sensor.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Industrial Safety Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { SensorStatus } from './sensorStatus';
import type { SensorType } from './sensorType';

export interface Sensor {
  id: string;
  name: string;
  sensorType: SensorType;
  zoneId: string;
  zoneName: string;
  unit: string;
  latestValue: number;
  status: SensorStatus;
  updatedAt: Date;
}