# lib/api-zod/src/generated/types/sensorStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Industrial Safety Intelligence System API
 * OpenAPI spec version: 0.1.0
 */

export type SensorStatus = typeof SensorStatus[keyof typeof SensorStatus];


export const SensorStatus = {
  normal: 'normal',
  warning: 'warning',
  critical: 'critical',
} as const;