# lib/api-zod/src/generated/types/sensorType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Industrial Safety Intelligence System API
 * OpenAPI spec version: 0.1.0
 */

export type SensorType = typeof SensorType[keyof typeof SensorType];


export const SensorType = {
  gas: 'gas',
  vibration: 'vibration',
  temperature: 'temperature',
  pressure: 'pressure',
  proximity: 'proximity',
  camera: 'camera',
} as const;