# lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Industrial Safety Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { TimelineEventEventType } from './timelineEventEventType';

export interface TimelineEvent {
  id: string;
  hazardId: string;
  hazardTitle: string;
  eventType: TimelineEventEventType;
  description: string;
  timestamp: Date;
}